+++ internal/onnxgomlx/fusion/quantized_qkv.go
@@ -2,9 +2,9 @@ package fusion
 
 import (
 	"github.com/gomlx/compute"
-	. "github.com/gomlx/gomlx/pkg/core/graph" //nolint
-	"github.com/gomlx/gomlx/pkg/ml/context"
-	"github.com/gomlx/gomlx/pkg/ml/nn"
+	. "github.com/gomlx/gomlx/core/graph" //nolint
+	"github.com/gomlx/gomlx/ml/model"
+	"github.com/gomlx/gomlx/ml/nn"
 	"github.com/gomlx/onnx-gomlx/internal/onnxgomlx"
 )
 
@@ -40,7 +40,7 @@ func (c *quantizedQKVDenseCandidate) OutputNames() []string {
 func (c *quantizedQKVDenseCandidate) InternalOutputs() map[string]bool { return c.internalOutputs }
 func (c *quantizedQKVDenseCandidate) ExternalInputs() []string         { return c.externalInputs }
 
-func (c *quantizedQKVDenseCandidate) Emit(_ *context.Context, g *Graph, convertedOutputs map[string]*Node) {
+func (c *quantizedQKVDenseCandidate) Emit(_ *model.Scope, g *Graph, convertedOutputs map[string]*Node) {
 	p := c.params
 
 	floatInput := convertedOutputs[p.FloatInputName]